Autopath module
Allows for tokenised url paths to be generated for an entity based on
text fields for that entity.

e.g a blog path that doesn't need a path written out for it each time
but is based on its title. Paths can simply be overriden by typing in a
value for the path field.

`[fieldName]` is swapped out for a url friendly version of that field if
the path is left blank.

An admin form in the structure menu allows for creating of token
substitutions and offers list of tokens based on textfields of that
entity type. Only shows entities if they have  a `path` field.