fix(form): Completion handler for form renderes
Was an issue that form getting data run before they were rendered by JSONForm.

New featuer:
If doing stuff on a form as soon as it is rendered, there will be a new function that can be used:

formComplete_formName()

Attach this to the window of the site on the front end.

I.e.

window.formComplete_SignInForm = function(){
    init()
}

Signed-off-by: Alex Bor <alexhbor@gmail.com>