Allow objects in field type definitions
Example:

```JSON

{
"name": "Image",
"type": {"title":"String","alt":"String","path":"String"}
}

```

Normal field form hooks run for entityUI but it also has a default of
simply showing an object with fields of the type given.

Currently only supports String and Number for the default form handler.
More complex ones can be added in `entityUI.js` if this is shown to be a
good solution.