Redirect trailing slashes in urls
[Fixes #240]

I looked at Google's advice and it seemed to indicate allowing both
wasn't amazing SEO so I've done a redirect. This does however mean we
can't have a trailing slash in paths on purpose. Is that OK? We're not
allowing directory listing and if we ever do we can always change this
rule.