Remove Windows-breaking example.js and add install CLI
`npm install -g irisjs` will now add `irisjs` to the command line path.

Currently this serves only one use `irisjs install sitename` which steps
through a list of config questions, installs irisjs locally and creates
a launch file with the settings the user put in.

There is room to add other command line options in the future.

Requires an npm install for the `prompt` library that makes asking
questions on the command line easier with defaults and such.