Response intercept hook
Allow responses from the server to be intercepted before they are sent
to the client. Was built for use in nojs forms system but has many wider
uses. Could possibly be used to replace the headertags system.

Going to tidy up a few things here and add docs but putting it up to
show it's being worked on.